+++ crates/mds-collector/src/hosts_up_checker.rs
@@ -3,20 +3,23 @@ use std::thread::{self, JoinHandle};
 use std::time::{Duration, Instant};
 
 use mds_ipinfo::LastKnownStatus;
+use mds_util::host_up::TimeoutSettings;
 use mds_util::prelude::is_host_up;
 
 pub(super) struct HostsUpChecker {
     time_since_last_run: Instant,
     check_cooldown_secs: u16,
     handle: Option<JoinHandle<Vec<(IpAddr, LastKnownStatus)>>>,
+    timeout_settings: TimeoutSettings,
 }
 
 impl HostsUpChecker {
-    pub(super) fn new(check_cooldown_secs: u16) -> Self {
+    pub(super) fn new(check_cooldown_secs: u16, timeout_settings: TimeoutSettings) -> Self {
         Self {
             time_since_last_run: Instant::now(),
             check_cooldown_secs,
             handle: None,
+            timeout_settings,
         }
     }
 
@@ -25,6 +28,7 @@ impl HostsUpChecker {
             return;
         }
         self.time_since_last_run = Instant::now();
+        let timeout_settings = self.timeout_settings;
         let h: JoinHandle<Vec<(IpAddr, LastKnownStatus)>> = thread::Builder::new()
             .name("host_up_checker".to_string())
             .spawn(move || {
@@ -32,7 +36,7 @@ impl HostsUpChecker {
                 for ip in host_ips {
                     match ip {
                         IpAddr::V4(ipv4_addr) => {
-                            let status = if is_host_up(ipv4_addr, None) {
+                            let status = if is_host_up(ipv4_addr, None, timeout_settings) {
                                 LastKnownStatus::Online
                             } else {
                                 LastKnownStatus::Offline
@@ -81,7 +85,7 @@ mod tests {
 
     #[test]
     fn test_checker_run_and_finish_with_mixed_ips() {
-        let mut checker = HostsUpChecker::new(0);
+        let mut checker = HostsUpChecker::new(0, TimeoutSettings::default());
         let ips = vec![
             IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)),
             IpAddr::V4(IP_TEST_NET_1_UNREACHABLE),